changed int to long long int and fixed logical mistake in MoveNodesFast

+++ RecoHGCal/TICL/plugins/TracksterLinkingbyLeiden.h
@@ -55,7 +55,7 @@ namespace ticl {
     edm::ESHandle<MagneticField> bfield_;
     edm::ESHandle<Propagator> propagator_;
 
-    double gamma_{1};  //resolution parameter of the algortihm. The higher the gamma, the more communities are yielded
+    long long int gamma_{1};  //resolution parameter of the algortihm. The higher the gamma, the more communities are yielded
     double theta_{1};  //parameter of the refinement step
 
     void leidenAlgorithm(TICLGraph &graph, Partition &partition, std::vector<Flat> &flatFinalPartition);
@@ -67,13 +67,13 @@ bool isAlgorithmDone(ticl::TICLGraph const &graph, ticl::Partition const &partit
 ticl::Partition &removeEmptyCommunities(ticl::Partition &partition);
 
 ticl::Partition &refinePartition(
-    ticl::TICLGraph const &graph, ticl::Partition &partition, ticl::Partition &singlePartition, double gamma, double theta);
+    ticl::TICLGraph const &graph, ticl::Partition &partition, ticl::Partition &singlePartition, long long int gamma, double theta);
 
-ticl::Partition &moveNodesFast(ticl::Partition &partition, double gamma);
+ticl::Partition &moveNodesFast(ticl::Partition &partition, long long int gamma);
 
 ticl::Partition &singletonPartition(ticl::TICLGraph const &graph, ticl::Partition &singlePartition);
 
-ticl::Partition &mergeNodesSubset(ticl::Partition &partition, ticl::Community const &subset, double gamma);
+ticl::Partition &mergeNodesSubset(ticl::Partition &partition, ticl::Community const &subset, long long int gamma);
 
 void aggregateGraph(ticl::TICLGraph &graph, ticl::Partition const &partition);
 #endif
